Move prepare retry logic from LiveReload to execCordovaPrepare.
Currently LiveReload will try a second time to prepare if the first time fails (usually because of errors trying to create www folder to quickly on Windows). This moves that logic into execCordovaPrepare itself, so we gain the same benefit in other scenarios.

(I'm seeing this failure regularly on the first launch of simulations from VS Code)