Improving XHR proxy
With the XHR proxy enabled, previously we would lose the static properties of XMLHttpRequest, such as DONE and LOADING. We now make sure to include these properties.

Additionally, since the connection from the app to the proxy always succeeds we do not correctly handle network level failures. Now if there is an error in the communication itself (as distinct from recieving a 404 for example) we report back a 502 Bad Gateway error, and include the error message to inform the app developer about the issue.