Make compass 'wrapper size' multiple of 2 (#167)
This determines how much bigger the background is compared to the compass. It needs to be a multiple of 2, since it is divided by 2 to determine the center point of the outer circle. At its previous value, the compass wasn't centered within that outer circle (12 pixels top and left, 13 pixels bottom and right).