Initial support for themes.
A themes can be specified via the --theme commandline switch or a theme property on the options API property. It can be a file name pointing to a JSON file, or a JavaScript object or JSON string. The format is as specified here: https://github.com/Microsoft/cordova-simulate/wiki/Support-for-Themes